Add logging functionality to brain.py.
For every try/except block, logging will have to be called so that the exception can be
logged for debugging. I changed the placeholder code in the changed files to reflect this.

Note: I might want to add different .log files for different parts of the project.

Example:
	Core project logging can be placed in unison.log
	Module logging can be placed in module.log
	Additional logging can be placed in yet another log file

This separates each log so that a developer can more easily debug the code from a user's,
contributors, and developers prospective.