Add quit_condition method to switch.py; add licence field to default.yml; delete commented copy/paste code.
The commented copy/paste code was deleted as to ensure proper security when developing the program. A choice has
been made to allow users to copy the programs .exe files they want to use with the program into the Unison Desktop
directory. This could be done automatically, but not at the risk of security.