Add generic implementation to internet_browser; restructure apis directory.
The generic implementation of chrome_browser to internet_browser makes it easier for a user
to choose which internet browser they would like to use. For example, if a user wants to use FireFox,
all they need to do is copy the "firefox.exe" (or the FireFox .exe file) path into their user profile
and associated it with the internet_browser key.

Example:
	name: John Smith
	email: johnsmith@mail.com
	internet_browser: C:/Program Files/Firefox/firefox.exe  <--

The restructure of the apis directory enables for more structured and stable releases in the future.